app/backend/requirements.txt
fastapi~=0.115.0
pydantic~=2.9.2
Faker~=30.3.0
uvicorn
passlib~=1.7.4
sqlalchemy~=2.0.35
python-jose